% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all.R
\name{experimental_results}
\alias{experimental_results}
\title{Report the results of the experiment as well as confidence intervals.}
\usage{
experimental_results(ys, xs, alpha = 0.05)
}
\arguments{
\item{ys}{The measurements of the response}

\item{xs}{The design}

\item{alpha}{\code{1 - alpha} is the confidence of the computed intervals. Default is \code{0.05}.}
}
\value{
A list of the estimate as well as confidence intervals
}
\description{
Report the results of the experiment as well as confidence intervals.
}
\author{
Adam Kapelner
}

