% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\docType{data}
\name{napth}
\alias{napth}
\title{This is data for the PRV measurement of the k_H of Napthalene in water.
See Section 3 of our paper below for more information.}
\format{
A data frame with 100 rows and 2 variables
}
\usage{
data(napth)
}
\description{
This is data for the PRV measurement of the k_H of Napthalene in water.
See Section 3 of our paper below for more information.
}
\references{
\url{https://arxiv.org/abs/1604.03480}
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{datasets}
