% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all.R
\name{oed_for_slope_over_intercept}
\alias{oed_for_slope_over_intercept}
\title{Create an optimal design for measuring the slope divided by the intercept}
\usage{
oed_for_slope_over_intercept(n, xmin, xmax, theta0, f_hetero = NULL, ...)
}
\arguments{
\item{n}{The number of experimental runs.}

\item{xmin}{The minimum value of the independent variable.}

\item{xmax}{The maximum value of the independent variable.}

\item{theta0}{The guess of the true value of the slope / intercept.}

\item{f_hetero}{Specification of heteroskedasticity: the h(x) which relates the value of the
                        independent variable to the variance in the response around the line at that place
                        or the proportional variance at that point. If \code{NULL}, homoskedasticity is
                        assumed (this is the default behavior).}
}
\value{
An n-vector of x-values which specifies the optimal design
}
\description{
Create an optimal design for measuring the slope divided by the intercept
}
\author{
Adam Kapelner
}

