% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/public.R
\name{experimental_results}
\alias{experimental_results}
\title{Report the results of the experiment as well as confidence intervals.}
\usage{
experimental_results(ys, xs, alpha = 0.05, B = 1000)
}
\arguments{
\item{ys}{The measurements of the response}

\item{xs}{The design}

\item{alpha}{\code{1 - alpha} is the confidence of the computed intervals. Default is \code{0.05}.}

\item{B}{For the confidence interval methods with an embedded bootstrap (or resampling), the number
                    of resamples (defaults to \code{1000}).}
}
\value{
A list of the estimate as well as confidence intervals
}
\description{
Report the results of the experiment as well as confidence intervals.
}
\author{
Adam Kapelner
}

