% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optDesignSlopeInt.R
\docType{package}
\name{optDesignSlopeInt}
\alias{optDesignSlopeInt}
\title{Optimal Designs for Estimating the Slope Divided by the Intercept}
\description{
Software which helps practitioners optimally design experiments that measure the slope divided by the intercept.
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{Design}
\keyword{Experiments}
\keyword{Optimality}
